Move logic related to filepaths to clearml/storage/filepaths.py
